Upstream.

+++ _ref/ops.sh
@@ -11,6 +11,8 @@ _convert_bash-backend-lowLatency() {
     jq -Rs '{ model: "meta-llama/llama-3.1-405b-instruct", provider: { "order": ["Lambda", "Fireworks"], "sort": "latency" }, messages: [{"role": "user", "content": .}] }' | curl -fsS --max-time 120 --keepalive-time 300 --compressed --tcp-fastopen --http2 -X POST https://openrouter.ai/api/v1/chat/completions -H "Content-Type: application/json" -H "Authorization: Bearer $OPENROUTER_API_KEY" --data-binary @- | jq -r '.choices[0].message.content'
 }
 _convert_bash-dispatch() {
+    [[ "$1" == "" ]] && return 1
+    [[ ! -e "$1" ]] && return 1
     echo 'quick brown fox' | _convert_bash-backend > /dev/null
     
     #-s 4096
@@ -47,6 +49,8 @@ _semanticAssist_bash-backend-lowLatency-special() {
 # ATTENTION: Override with 'ops.sh' or similar if appropriate.
 # (ie. usually to change parallelization for high-latency APIs, providers, etc)
 _semanticAssist_bash-dispatch() {
+    [[ "$1" == "" ]] && return 1
+    [[ ! -e "$1" ]] && return 1
     echo 'quick brown fox' | _semanticAssist_bash-backend > /dev/null
     
     #-s 4096
